Got tests passing, updated imports and moved a couple of classes
