+++ opencsp/common/lib/cv/spot_analysis/image_processor/test/test_AbstractSpotAnalysisImageProcessor.py
@@ -6,24 +6,10 @@ import unittest
 import numpy as np
 
 import opencsp.common.lib.cv.CacheableImage as ci
-from opencsp.common.lib.cv.spot_analysis.image_processor import AbstractSpotAnalysisImageProcessor
 import opencsp.common.lib.cv.spot_analysis.SpotAnalysisOperable as sao
 import opencsp.common.lib.tool.file_tools as ft
 import opencsp.common.lib.tool.log_tools as lt
-
-
-class DoNothingImageProcessor(AbstractSpotAnalysisImageProcessor):
-    def _execute(self, operable: sao.SpotAnalysisOperable, is_last: bool) -> list[sao.SpotAnalysisOperable]:
-        return [operable]
-
-
-class SetOnesImageProcessor(AbstractSpotAnalysisImageProcessor):
-    def _execute(self, operable: sao.SpotAnalysisOperable, is_last: bool) -> list[sao.SpotAnalysisOperable]:
-        img = copy.copy(operable.primary_image.nparray)
-        img[:, :] = 1
-        ret = dataclasses.replace(operable, primary_image=img)
-        return [ret]
-
+from opencsp.common.lib.cv.spot_analysis.image_processor.AbstractSpotAnalysisImageProcessor import DoNothingImageProcessor, SetOnesImageProcessor
 
 class test_AbstractSpotAnalysisImageProcessor(unittest.TestCase):
     @classmethod
